percent fields to float fields

+++ lastwill/contracts/submodels/swaps.py
@@ -178,8 +178,8 @@ class ContractDetailsSWAPS2(CommonDetails):
 
     broker_fee = models.BooleanField(default=False)
     broker_fee_address = models.CharField(max_length=50, null=True, default=None)
-    broker_fee_base = models.IntegerField(null=True, default=None)
-    broker_fee_quote = models.IntegerField(null=True, default=None)
+    broker_fee_base = models.FloatField(null=True, default=None)
+    broker_fee_quote = models.FloatField(null=True, default=None)
 
     eth_contract = models.ForeignKey(
         EthContract,
